Provide new function for starting monerod and monerowalletrpc
