+++ src/monerod.rs
@@ -54,43 +54,3 @@ struct GetBlockHeaderByHeight {
     status: String,
     untrusted: bool,
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::Duration;
-    use testcontainers::images::generic::{Stream, WaitFor};
-    use testcontainers::{clients, core::Port, images, Docker, Image};
-
-    #[tokio::test]
-    async fn can_get_genesis_block_header() {
-        let docker = clients::Cli::default();
-
-        let image = images::generic::GenericImage::new("xmrto/monero")
-            .with_mapped_port(Port {
-                local: 28081,
-                internal: 28081,
-            })
-            .with_mapped_port(Port {
-                local: 28083,
-                internal: 28083,
-            })
-            .with_entrypoint("")
-            .with_args(vec![
-                "/bin/bash".to_string(),
-                "-c".to_string(),
-                "monerod --confirm-external-bind --non-interactive --regtest --rpc-bind-ip 0.0.0.0 --rpc-bind-port 28081 --no-igd --hide-my-port --fixed-difficulty 1 --rpc-payment-allow-free-loopback --data-dir /monero & \
-                monero-wallet-rpc --log-level 4 --daemon-address localhost:28081 --confirm-external-bind --rpc-login username:password --rpc-bind-ip 0.0.0.0 --rpc-bind-port 28083 --daemon-login username:password --wallet-dir /monero/".to_string(),
-            ])
-            .with_wait_for(WaitFor::LogMessage { message: "You are now synchronized with the network. You may now start monero-wallet-cli".to_string(), stream: Stream::StdOut });
-        let node = docker.run(image);
-
-        let cli = Client::localhost(28081).unwrap();
-        let value = cli
-            .get_block_header_by_height(0)
-            .await
-            .expect("failed to get block 0");
-
-        assert_eq!(0, value.height);
-    }
-}
